requirements.txt
pygame==2.1.1
yandex-music==0.18.0
Pillow==8.4.0
